fix: dark theme for comments screen
