+++ lib/screens/timeline_screen/all_event_screen.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:hng_events_app/screens/comment_screen.dart';
 import 'package:hng_events_app/screens/timeline_screen/my_events_screen.dart';
+import 'package:hng_events_app/util/date_formatter.dart';
 import 'package:hng_events_app/widgets/timeline_event_card.dart';
 
 import '../../constants/colors.dart';
@@ -66,7 +67,7 @@ class AllEventsScreen extends ConsumerWidget {
                   eventNotifier.allEvents?.data.events[index].location ?? "N/A",
               date: eventNotifier.allEvents?.data.events[index].startDate ??
                   "N/A",
-              activity: timeLeft(
+              activity: DateFormatter().timeLeft(
                   eventNotifier.allEvents!.data.events[index].startDate,
                   eventNotifier.allEvents!.data.events[index].startTime),
             ),
@@ -76,30 +77,6 @@ class AllEventsScreen extends ConsumerWidget {
     );
   }
 
-  // static String timeLeft(DateTime date) {
-  //   final date2 = DateTime.now();
-  //   final difference = date.difference(date2);
-
-  //   if ((difference.inDays / 7).floor() >= 1) {
-  //     return '1 week Left';
-  //   } else if (difference.inDays >= 2) {
-  //     return '${difference.inDays} days Left';
-  //   } else if (difference.inDays >= 1) {
-  //     return '1 day Left';
-  //   } else if (difference.inHours >= 2) {
-  //     return '${difference.inHours} hours Left';
-  //   } else if (difference.inHours >= 1) {
-  //     return '1 hour Left';
-  //   } else if (difference.inMinutes >= 2) {
-  //     return '${difference.inMinutes} minutes Left';
-  //   } else if (difference.inMinutes >= 1) {
-  //     return '1 minute Left';
-  //   } else if (difference.inSeconds >= 3) {
-  //     return '${difference.inSeconds} seconds Left';
-  //   } else {
-  //     return 'Expired';
-  //   }
-  // }
 }
 
 Widget eventCard(
