+++ lib/screens/timeline_screen/my_events_screen.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:hng_events_app/screens/comment_screen.dart';
+import 'package:hng_events_app/util/date_formatter.dart';
 import 'package:hng_events_app/widgets/timeline_event_card.dart';
 import 'package:intl/intl.dart';
 
@@ -68,50 +69,12 @@ class MyEventScreen extends ConsumerWidget {
                 time: event.startTime,
                 location: event.location,
                 date: event.startDate,
-                activity: timeLeft(event.startDate, event.startTime),
+                activity: DateFormatter().timeLeft(event.startDate, event.startTime),
               ),
             );
           },
         ),
       ),
     );
-  }
-}
-
-String timeLeft(String startDate, String startTime) {
-  //final formatter =DateFormat('HH:mm');
-  final DateTime date;
-
-  if (RegExp(r'^[0-9]{2}:[0-9]$').hasMatch(startTime)) {
-    startTime = '${startTime.substring(0, 3)}0${startTime.substring(3)}';
-    date = DateTime.parse("${startDate}T$startTime");
-  } else if (RegExp('^[0-9]:[0-9]{2}').hasMatch(startTime)) {
-    startTime = '0${startTime.substring(0)}';
-    date = DateTime.parse("${startDate}T$startTime");
-  } else {
-    date = DateTime.parse("${startDate}T$startTime");
-  }
-
-  final date2 = DateTime.now();
-  final difference = date.difference(date2);
-
-  if ((difference.inDays / 7).floor() >= 1) {
-    return '1 week Left';
-  } else if (difference.inDays >= 2) {
-    return '${difference.inDays} days Left';
-  } else if (difference.inDays >= 1) {
-    return '1 day Left';
-  } else if (difference.inHours >= 2) {
-    return '${difference.inHours} hours Left';
-  } else if (difference.inHours >= 1) {
-    return '1 hour Left';
-  } else if (difference.inMinutes >= 2) {
-    return '${difference.inMinutes} minutes Left';
-  } else if (difference.inMinutes >= 1) {
-    return '1 minute Left';
-  } else if (difference.inSeconds >= 3) {
-    return '${difference.inSeconds} seconds Left';
-  } else {
-    return 'Expired';
   }
 }
